[Effector] select method with config class
